requirements/dev.txt
astroid==1.5.3
certifi==2017.11.5
chardet==3.0.4
codecov==2.0.10
colorama==0.3.9
Django==2.0
django-rest-framework==0.1.0
djangorestframework==3.7.7
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.7.4
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
wrapt==1.10.11
